refactor: extract WidgetTokensUpdater
