+++ apps/cowswap-frontend/src/entities/balancesContext/balancesContextAtom.ts
@@ -1,0 +1,7 @@
+import { atom } from 'jotai'
+
+export interface BalancesContext {
+  account: string | null
+}
+
+export const balancesContextAtom = atom<BalancesContext>({ account: null })
